Add files via upload
Led the development of an engaging and user-friendly web portal for ASV Empowerment to champion the journey of neurodivergent individuals towards self-actualization. The portal, crafted with a keen understanding of neurodiverse needs, offers a non-intimidating environment with resources that support the community in their interactions with a neurotypical society. The website features a suite of tailored services including a free journal download to engage visitors, membership subscriptions to exclusive

content, and premium courses tailored to various aspects of neurodivergent living. I was responsible for integrating e-commerce functionality for seamless transactions, as well as implementing interactive community features and monthly group coaching sessions. My role ensured a balance between aesthetic warmth and cognitive accessibility, resulting in a platform that empowers neurodivergent adults to thrive without masking their true selves.